refactor(editor): unify simulation expressions and result presentation
Split setup editing and diagnostic components from workspace orchestration. Retire the unused handwritten panel and deprecated Probe type. Plot derived scalar/complex series directly, hydrate bounded receipts from File artifacts, restore setup-local results, and keep session failures recoverable. Use canonical Profile defaults and asynchronous artifact archives. Preserve OP canvas projection, comparisons, measurements and plot exports.

Validation: 230 focused contracts pass, including output/plot and browser-session units; typecheck passes. Added browser regression for setup switching without rerunning. Full delivery selected by shared Model contract and lockfile, to be run on this integrated branch.

Test-Impact: tests-updated
