+++ apps/editor/src/features/simulation/dc-results-explorer.test.tsx
@@ -31,7 +31,7 @@ describe("DcResultsExplorer", () => {
         probes={[
           {
             id: "probe-out",
-            kind: "net-voltage",
+            kind: "voltage",
             documentId: "tb",
             anchor: { kind: "base-net", netId: "out" },
             occurrence: [],
