+++ apps/editor/src/features/simulation/simulation-focus-target.ts
@@ -1,0 +1,7 @@
+import type { SimulationExpression } from "@icm/model";
+
+/** A chart output id plus its canonical Canvas expression, not a stored probe protocol. */
+export type SimulationFocusTarget = { id: string } & Extract<
+  SimulationExpression,
+  { kind: "voltage" | "current" }
+>;
