+++ apps/editor/src/features/simulation/simulation-output-results.tsx
@@ -1,37 +1,23 @@
+import type { SimulationFocusTarget } from "./simulation-focus-target";
 import {
   simulationExpressionDependencies,
   type SimulationOutputSpec,
-  type SimulationProbeSpec,
 } from "@icm/model";
-import type { AcResult, TransientResult } from "@icm/spice-run";
-import type {
-  Prepared,
-  SimulationOutputData,
-} from "@icm/simulation-service/contract";
+import type { SimulationOutputData } from "@icm/simulation-service/contract";
 
-import { AcResultsExplorer } from "./ac-results-explorer";
-import { TransientResultsExplorer } from "./transient-results-explorer";
+import {
+  ComplexResultsExplorer,
+  unwrapPhaseDegrees,
+} from "./ac-results-explorer";
+import { ScalarResultsExplorer } from "./transient-results-explorer";
 import { SimulationMeasurementResults } from "./simulation-measurement-results";
 
-function focusProbe(output: SimulationOutputSpec): SimulationProbeSpec | null {
+function focusProbe(
+  output: SimulationOutputSpec,
+): SimulationFocusTarget | null {
   const dependency = simulationExpressionDependencies(output.expression)[0];
   if (!dependency) return null;
-  return dependency.kind === "voltage"
-    ? {
-        id: output.id,
-        kind: "net-voltage",
-        documentId: dependency.documentId,
-        anchor: structuredClone(dependency.anchor),
-        occurrence: [...dependency.occurrence],
-      }
-    : {
-        id: output.id,
-        kind: "terminal-current",
-        documentId: dependency.documentId,
-        instanceId: dependency.instanceId,
-        pinName: dependency.pinName,
-        occurrence: [...dependency.occurrence],
-      };
+  return { ...dependency, id: output.id };
 }
 
 export function SimulationOutputResults({
@@ -43,18 +29,13 @@ export function SimulationOutputResults({
   resultKey: string;
   data: SimulationOutputData;
   outputs: readonly SimulationOutputSpec[];
-  onFocusProbe?(probe: SimulationProbeSpec): void;
+  onFocusProbe?(probe: SimulationFocusTarget): void;
 }) {
   const authored = new Map(outputs.map((output) => [output.id, output]));
   const probes = outputs.flatMap((output) => {
     const probe = focusProbe(output);
     return probe ? [probe] : [];
   });
-  const labels = Object.fromEntries(
-    outputs.map((output) => [output.id, output.label]),
-  );
-  const vectors = (ids: readonly string[]): Prepared["vectors"] =>
-    ids.map((id) => ({ probeId: id, vector: id, quantity: "voltage" }));
   const visibleAnalyses = new Set(
     data.analyses.map((analysis) => analysis.analysis),
   );
@@ -102,61 +83,62 @@ export function SimulationOutputResults({
         return (
           <section key={`${analysis.analysis}-${analysisIndex}`}>
             {analysis.analysis === "ac" && complex.length > 0 ? (
-              <AcResultsExplorer
+              <ComplexResultsExplorer
                 resultKey={`${resultKey}:complex:${analysisIndex}`}
-                analysis={
-                  {
-                    analysis: "ac",
-                    plotName: analysis.plotName,
-                    frequencyHz: analysis.domain.values,
-                    probes: complex.map((output) => ({
-                      name: output.id,
-                      quantity: "derived",
-                      unit: output.unit,
-                      real: output.values.map((value) => value ?? Number.NaN),
-                      imag: output.imaginary!.map(
-                        (value) => value ?? Number.NaN,
-                      ),
+                plotName={analysis.plotName}
+                traces={complex.map((output, colorIndex) => {
+                  const phases = unwrapPhaseDegrees(
+                    output.values.map(
+                      (value, i) =>
+                        (Math.atan2(
+                          output.imaginary![i] ?? Number.NaN,
+                          value ?? Number.NaN,
+                        ) *
+                          180) /
+                        Math.PI,
+                    ),
+                  );
+                  const probe = probes.find((probe) => probe.id === output.id);
+                  return {
+                    id: output.id,
+                    label: output.label,
+                    colorIndex,
+                    quantity:
+                      output.unit === "V"
+                        ? "voltage"
+                        : output.unit === "A"
+                          ? "current"
+                          : output.unit === "1"
+                            ? "ratio"
+                            : output.unit,
+                    ...(probe ? { probe } : {}),
+                    points: analysis.domain!.values.map((frequency, i) => ({
+                      frequency,
+                      magnitudeDb:
+                        20 *
+                        Math.log10(
+                          Math.max(
+                            Math.hypot(
+                              output.values[i] ?? Number.NaN,
+                              output.imaginary![i] ?? Number.NaN,
+                            ),
+                            1e-30,
+                          ),
+                        ),
+                      phaseDeg: phases[i] ?? Number.NaN,
                     })),
-                  } satisfies AcResult
-                }
-                vectors={vectors(complex.map((output) => output.id))}
-                probes={probes.filter((probe) =>
-                  complex.some((output) => output.id === probe.id),
-                )}
-                labels={labels}
-                groups={Object.fromEntries(
-                  complex.map((output) => [
-                    output.id,
-                    output.unit === "V"
-                      ? "voltage"
-                      : output.unit === "A"
-                        ? "current"
-                        : output.unit === "1"
-                          ? "ratio"
-                          : output.unit,
-                  ]),
-                )}
+                  };
+                })}
                 {...(onFocusProbe ? { onFocusProbe } : {})}
               />
             ) : null}
             {[...scalarByUnit.entries()].map(([unit, unitOutputs]) => {
-              const synthetic: TransientResult = {
-                analysis: "tran",
-                plotName: analysis.plotName,
-                timeSeconds: analysis.domain!.values,
-                probes: unitOutputs.map((output) => ({
-                  name: output.id,
-                  quantity: "derived",
-                  unit: unit === "1" ? null : unit,
-                  value: output.values.map((value) => value ?? Number.NaN),
-                })),
-              };
               return (
-                <TransientResultsExplorer
+                <ScalarResultsExplorer
                   key={unit}
                   resultKey={`${resultKey}:${analysis.analysis}:${analysisIndex}:${unit}`}
-                  analysis={synthetic}
+                  plotName={analysis.plotName}
+                  domain={analysis.domain!.values}
                   analysisLabel={
                     analysis.analysis === "tran"
                       ? "Transient"
@@ -165,10 +147,11 @@ export function SimulationOutputResults({
                   domainLabel={analysis.domain!.name}
                   domainUnit={analysis.domain!.unit}
                   logarithmicX={analysis.analysis === "ac"}
-                  vectors={vectors(unitOutputs.map((output) => output.id))}
-                  groups={Object.fromEntries(
-                    unitOutputs.map((output) => [
-                      output.id,
+                  traces={unitOutputs.map((output, colorIndex) => ({
+                    id: output.id,
+                    label: output.label,
+                    colorIndex,
+                    quantity:
                       unit === "V"
                         ? "voltage"
                         : unit === "A"
@@ -176,12 +159,16 @@ export function SimulationOutputResults({
                           : unit === "1"
                             ? "Value"
                             : unit,
-                    ]),
-                  )}
-                  probes={probes.filter((probe) =>
-                    unitOutputs.some((output) => output.id === probe.id),
-                  )}
-                  labels={labels}
+                    unit: unit === "1" ? null : unit,
+                    values: output.values.map((value) => value ?? Number.NaN),
+                    ...(probes.find((probe) => probe.id === output.id)
+                      ? {
+                          probe: probes.find(
+                            (probe) => probe.id === output.id,
+                          )!,
+                        }
+                      : {}),
+                  }))}
                   {...(onFocusProbe ? { onFocusProbe } : {})}
                 />
               );
