+++ docs/specs/simulation.md
@@ -2,8 +2,9 @@
 
 Status: accepted
 
-Owners: `packages/spice-run`, `apps/local-host`, `worker`,
-`containers/ngspice`
+Owners: `packages/model`, `packages/netlist`, `packages/simulation-service`,
+`apps/editor/src/features/simulation`, `apps/local-host`, `worker`,
+`packages/spice-run`, `containers/ngspice`
 
 Related decision: [ADR 0055](../adr/0055-simulation-is-part-of-the-product.md)
 
@@ -14,8 +15,7 @@ testbench intent and analysis commands. The execution environment supplies the
 simulator-readable model libraries. These are separate responsibilities and
 are assembled into simulation input without changing circuit Documents or the
 structural netlist export. Authored simulation intent may be saved only through
-the optional Project `SimulationSetup` defined below; that field moves the
-Project schema when it lands.
+the Project `simulationSetups` collection defined below.
 
 This contract covers deck assembly, model-library selection, and, since the
 2026-09-04 amendment of ADR 0055, the shape of a run: its inputs, root,
@@ -66,12 +66,12 @@ At container startup the harness measures the real binary, model tree,
 platform, and `.spiceinit` bytes. It reports `reproducibility: "pinned"` and the
 Profile ID only when every measured identity matches. A missing, malformed, or
 mismatched Profile leaves `/health` and `/run` not ready; it is never silently
-downgraded to an observed hosted run. An ordinary local host remains
-`observed` and may use its explicitly configured library without claiming this
-Profile.
+downgraded to an observed hosted run. Local deployment is an optional adapter,
+not an implicit fallback to a program found on the user's machine.
 
 The first qualified scope is deliberately narrow and factual: the continuous
-`sky130_fd_pr__nfet_01v8` and `sky130_fd_pr__pfet_01v8` wrappers, `tt`, and
+`sky130_fd_pr__nfet_01v8` and `sky130_fd_pr__pfet_01v8` wrappers, the
+Profile's qualified sections (`tt/ff/ss/fs/sf`), and
 OP/DC/AC/TRAN covered by the hosted model acceptance fixture. A separate
 independent-source divider smoke gives DC parsing and numerical sweep behavior
 an exact closed-form check on the same pinned runtime. TRAN qualification includes an ideal RC
@@ -153,15 +153,42 @@ still covers its exact emitted spelling.
 An environment fingerprint is the SHA-256 of canonical environment facts. A
 consumer verifies that fingerprint before accepting a runner response. The
 hosted container reports `pinned` only after matching the Profile at startup;
-the local host reports `observed` facts from the user's installation with null
-Profile and startup identities. Metadata plumbing alone never claims
-reproducibility.
+an explicitly configured local adapter must report its actual environment,
+using `observed` unless it satisfies a pinned Profile. The unconfigured local
+host produces no simulation metadata. Metadata plumbing alone never claims reproducibility.
 
 This envelope establishes provenance only. The numbers themselves are
 [Result data](#result-data). Bindings from a probe back to circuit objects
 are produced at compile time and are not part of reading a rawfile.
 
 ## Ownership boundary
+
+The implementation has these boundaries:
+
+- Model owns saved setups and canonical voltage/current expressions.
+- Netlist compiles structured intent. The service's preparation module resolves
+  raw/structured inputs, capabilities and immutable deck identity.
+- SimulationService owns run lifecycle, idempotency and retention. Executor is
+  its execution port; GUI and MCP use the same service and File Resource.
+- spice-run separates request/result types, deck assembly, metadata and terminal
+  verdicts. Its public exports remain the same.
+- GUI setup editing, diagnostic display and result materialization are separate
+  from workspace orchestration. Charts consume scalar/complex series; derived
+  outputs are not re-encoded as fake simulator analyses.
+
+The local host has no default binary, model directory or version, and does not
+probe installed software. Its optional `simulationHandler` uses the same
+`/api/simulate` transport as the hosted executor. Without it, capabilities
+reports `configured: false`, and execution returns
+`simulation-not-configured`; Editor and saved inputs remain usable.
+
+Run reads return bounded receipts. Full `result.json` and `outputs.json`
+remain File artifacts; the GUI materializes these only when a receipt is a
+preview. MCP keeps paged file access instead of receiving unbounded arrays.
+Input freshness is cached per Project structure revision and setup ID; raw
+dependency declarations participate in identity, not a separate unavailable state.
+Switching setups restores their session-local results; these are not saved in
+the Project. Lost responses never trigger an automatic second execution.
 
 - The circuit netlist is generated by `@icm/netlist`.
 - Model-library selection belongs to the local or hosted execution
@@ -183,9 +210,8 @@ and contains exactly one input form:
   selection; or
 - a raw setup with its entry path, authored files, and declared dependencies.
 
-Adding that optional field requires one Project schema migration. Existing
-Projects migrate with the field absent and retain their existing circuit,
-hierarchy, and structural-export behavior. The setup follows the ordinary
+Projects without authored setups have an empty collection and retain their
+existing circuit, hierarchy, and structural-export behavior. The setup follows the ordinary
 Project save, recovery, Gallery, revision, and undo/redo boundaries; it is not
 stored in a simulation-only sidecar or a second persistence service.
